+++ packages/hooks/src/index.js
@@ -6,11 +6,14 @@ const { createHookContext } = require('./create-hook-context');
 const { getHook } = require('./create-hooks-registry');
 const constants = require('./constants');
 
+const runHook = (...args) => createHook(...args);
+
 module.exports = {
   traceHook,
   logTrace,
   logBoot,
   createHook,
+  runHook,
   registerAction,
   createHookApp,
   runHookApp,