+++ packages/hooks/CHANGELOG.md
@@ -6,3 +6,4 @@
 - Features and services names defaults are derived by their registration form and automatically prefixed
 - Added support for positional definition in `registerRoute`
 - Allows for order-independent service2service and feature2feature integration
+- Deprecate "createHook" in favor of "runHook"